pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "grandata"
version = "0.1"
description = "CREsted-style genome-track data management backed by XArray."
readme = { file = "README.md", "content-type" = "text/markdown" }
requires-python = ">=3.10"

classifiers = [
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]

authors = [
  { name = "Matthew Schmitz" }
]

dependencies = [
    # "anndata",
    # "crested",
    "xarray",
    #"torch",
    "torchdata",
    "zarr>=3.0.0",
    "pybigtools",
    "pandas",
    "numpy",
    "pybigtools",
    "tqdm",
    "ete3",
    "pyfaidx"
]

[project.urls]
"Source" = "https://github.com/mtvector/grandata"

[tool.setuptools]
packages = { find = {} }

# [tool.setuptools.packages.find]
# exclude = ["tests*"]
